refactor: InputForm내 컴포넌트 분리, 함수 분리
