+++ client/src/component/InputForm/PaymentInput/PaymentInput.js
@@ -5,6 +5,7 @@ export default class PaymentInput {
     this.$target = $target; // $inputForm
     this.$paymentInput = document.createElement('div');
 
+    selectedHistoryStore.subscribe(() => this.render());
     paymentStore.subscribe(() => this.render());
 
     this.$target.appendChild(this.$paymentInput);
@@ -12,7 +13,48 @@ export default class PaymentInput {
     this.init();
   }
 
-  init() {}
+  init() {
+    this.$paymentInput.addEventListener('click', this.onClickPaymentItem);
+    this.$paymentInput.addEventListener('click', this.onClickDropdownField);
+  }
+
+  onClickPaymentItem(event) {
+    const $li = event.target.closest('.inputForm .payment>li');
+    if (!$li) return;
+
+    const $inputType = document.querySelector('input[name="payment"]');
+
+    if (!$li.dataset.name) {
+      // 추가하기 버튼
+      return;
+    }
+    $inputType.value = $li.dataset.name;
+    $inputType.dataset.id = $li.dataset.id;
+
+    const $payment = document.querySelector('.inputForm .payment');
+    $payment.style.display = 'none';
+  }
+
+  onClickDropdownField(event) {
+    const $field = event.target.closest('.field');
+    if (!$field) return;
+
+    const $dropdown = $field.nextElementSibling;
+    if (!$dropdown || !$dropdown.className.includes('dropdown')) return;
+
+    $dropdown.style.display =
+      ($dropdown.style.display === 'none') | ($dropdown.style.display === '')
+        ? 'block'
+        : 'none';
+
+    if (!$field) return;
+    const $categoryDropdown = $field.nextElementSibling.className.includes(
+      'category',
+    )
+      ? $field.nextElementSibling
+      : null;
+    if (!$categoryDropdown) return;
+  }
 
   render() {
     const history = selectedHistoryStore.get();
